UGENE-2495 fix
git-svn-id: https://local.ugene.unipro.ru/svn/ugene/trunk@5529 7ca3a678-163c-e011-b84f-001e8c0267ff

+++ src/corelibs/U2View/src/ov_phyltree/TreeViewer.h
@@ -289,7 +289,7 @@ private:
     
     PhyTreeObject*      phyObject;
     GraphicsBranchItem* root;
-    int                 maxNameWidth;
+    qreal               maxNameWidth;
     qreal               verticalScale;
     qreal               horizontalScale;
     qreal               view_scale;
